backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.9.9
djangorestframework>=3
djangorestframework-simplejwt>=5
django-cors-headers>=3.13.0
requests>=2.31.0
django-environ>=0.11.2
django-crontab>=0.7.1
joblib==1.2.0
numpy==1.23.4
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
django_pandas==0.6.7
matplotlib==3.9.1
mpld3==0.5.10